Solution to Check Equal To

+++ Check Equal To@@67882a731b4392d8fc1577da/code.c
@@ -1,9 +1,9 @@
 // Your code here...#include <stdio.h>
 
 int main() {
-    float num1, num2;
-    scanf("%f", &num1);
-    scanf("%f", &num2);
+    int num1, num2;
+    scanf("%d", &num1);
+    scanf("%d", &num2);
     if (num1 == num2) {
         printf("True\n");
     } else {
